Create AWS provider module
This commit creates the concept of providers. Using the existing
structures such as Stack, Connection Manager and Template we've created
Abstract Base Classes so that other providers can implement actions in
the future.

Future commits will deal with making the ABCs more generic for other
providers.
