+++ docs/_source/docs/hooks.rst
@@ -144,9 +144,9 @@ custom_hook.py
         argument: str
             The argument is available from the base class and contains the
             argument defined in the Sceptre config file (see below)
-        stack: sceptre.providers.stack.Stack
+        stack: sceptre.providers.aws.stack.Stack
              The associated stack of the hook.
-        connection_manager: sceptre.providers.connection_manager.ConnectionManager
+        connection_manager: sceptre.providers.aws.connection_manager.ConnectionManager
             Boto3 Connection Manager - can be used to call boto3 api.
 
         """
