+++ sceptre/providers/aws/stack_policies/lock.json
